+++ public/js/models/nics.js
@@ -1,13 +1,10 @@
 var Backbone = require('backbone');
 var _ = require('underscore');
+var Collection = require('./collection');
 
 
-var Nics = Backbone.Collection.extend({
+var Nics = Collection.extend({
     url: '/_/nics',
-    initialize: function(options) {
-        this.belongs_to_type = options.belongs_to_type;
-        this.belongs_to_uuid = options.belongs_to_uuid;
-    },
     mergeSysinfo: function(sysinfo) {
         _mergeSysinfo(sysinfo['Network Interfaces'], this, 'nic');
         _mergeSysinfo(sysinfo['Virtual Network Interfaces'], this, 'vnic');
@@ -30,22 +27,6 @@ var Nics = Backbone.Collection.extend({
                 nic.set(params);
             });
         }
-    },
-
-    fetchNics: function() {
-        var query = {};
-
-        if (this.belongs_to_uuid) {
-            query.belongs_to_uuid = this.belongs_to_uuid;
-        }
-
-        if (this.belongs_to_type) {
-            query.belongs_to_type = this.belongs_to_type;
-        }
-
-        this.fetch({
-            data: query
-        });
     }
 });
 module.exports = Nics;