+++ public/js/views/nics-row.js
@@ -1,11 +1,27 @@
 var Backbone = require('backbone');
+var adminui = require('../adminui');
+var NetworksDetailView = require('./networks-detail');
 
 var NicsRowView = Backbone.Marionette.ItemView.extend({
     tagName: 'tr',
     template: require('../tpl/nics-row.hbs'),
 
     events: {
-        'change input': 'onSelect'
+        'change input': 'onSelect',
+        'click .network-name': 'gotoNetwork'
+    },
+
+    gotoNetwork: function() {
+        var view = new NetworksDetailView({model: this.network});
+        view.render().$el.modal();
+    },
+
+    serializeData: function() {
+        var data = this.model.toJSON();
+        if (this.network) {
+            data.network_name = this.network.get('name');
+        }
+        return data;
     },
 
     onSelect: function(e) {
@@ -19,6 +35,6 @@ var NicsRowView = Backbone.Marionette.ItemView.extend({
             this.trigger('deselect', this.model);
         }
     }
-
 });
+
 module.exports = NicsRowView;